India vs Finland Comparison
With a little amount of collected data, the data analysis is done by using excel files to save the collected data and python is used to analyze this data. The aim of the project is to see if there is a correlation between deforestation, population and co2 emission which has a relationship with global warming. Two countries from completely distinct regions are used for comparison, India and Finland.